Don't change {remote,local}_settings.max_stream_id

+++ lib/includes/ngtcp2/ngtcp2.h
@@ -460,7 +460,7 @@ typedef struct {
 typedef struct {
   uint32_t max_stream_data;
   uint32_t max_data;
-  uint64_t max_stream_id;
+  uint32_t max_stream_id;
   uint16_t idle_timeout;
   uint8_t omit_connection_id;
   uint16_t max_packet_size;
